tools/requirements.txt
tweepy==3.9.0
python-dotenv==0.15.0
tqdm==4.51.0